ENH: Move to multi-command argparse setup
