+++ src/fmu_settings_cli/settings/__init__.py
@@ -1,0 +1,5 @@
+"""The 'settings' command."""
+
+from .main import CMD, add_parser, run
+
+__all__ = ["add_parser", "run", "CMD"]
