+++ src/fmu_settings_cli/settings/constants.py
