+++ tests/test_settings/test_start_api_server.py
@@ -4,14 +4,14 @@ from unittest.mock import patch
 
 import pytest
 
-from fmu_settings_cli.__main__ import generate_auth_token
-from fmu_settings_cli.api_server import start_api_server
+from fmu_settings_cli.settings.api_server import start_api_server
+from fmu_settings_cli.settings.main import generate_auth_token
 
 
 def test_start_api_server() -> None:
     """Tests that start_api_server calls as expected."""
     token = generate_auth_token()
-    with patch("fmu_settings_cli.api_server.run_server") as mock_run_server:
+    with patch("fmu_settings_cli.settings.api_server.run_server") as mock_run_server:
         start_api_server(token)
         mock_run_server.assert_called_once()
 
@@ -21,7 +21,8 @@ def test_start_api_server_fails() -> None:
     token = generate_auth_token()
     with (
         patch(
-            "fmu_settings_cli.api_server.run_server", side_effect=OSError("fail")
+            "fmu_settings_cli.settings.api_server.run_server",
+            side_effect=OSError("fail"),
         ) as mock_run_server,
         pytest.raises(RuntimeError, match="Could not start API server: fail"),
     ):
