+++ tests/test_settings/test_start_gui_server.py
@@ -5,14 +5,14 @@ from unittest.mock import patch
 import pytest
 from pytest import CaptureFixture
 
-from fmu_settings_cli.__main__ import generate_auth_token
-from fmu_settings_cli.gui_server import start_gui_server
+from fmu_settings_cli.settings.gui_server import start_gui_server
+from fmu_settings_cli.settings.main import generate_auth_token
 
 
 def test_start_gui_server() -> None:
     """Tests that start_gui_server calls as expected."""
     token = generate_auth_token()
-    with patch("fmu_settings_cli.gui_server.run_server") as mock_run_server:
+    with patch("fmu_settings_cli.settings.gui_server.run_server") as mock_run_server:
         start_gui_server(token)
         mock_run_server.assert_called_once()
 
@@ -22,7 +22,8 @@ def test_start_gui_server_fails() -> None:
     token = generate_auth_token()
     with (
         patch(
-            "fmu_settings_cli.gui_server.run_server", side_effect=OSError("fail")
+            "fmu_settings_cli.settings.gui_server.run_server",
+            side_effect=OSError("fail"),
         ) as mock_run_server,
         pytest.raises(RuntimeError, match="Could not start GUI server: fail"),
     ):
